Commit #6 - 01.09.2024
Added additional configuration to the Notification entity, and an additional virtual property to NotificationType. Added service and controller for notifications. Created a partial view for the notification form, which is used in the Index page. Created jQuery scripts for GET and POST requests to the relevant methods in the code-behind.